requirements.txt
asgiref==3.7.2
boto3==1.28.9
botocore==1.31.9
dj-database-url==0.5.0
Django==3.2.20
django-resized==1.0.2
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
Pillow==10.0.0
psycopg2==2.9.6
s3transfer==0.6.1
sqlparse==0.4.4
urllib3==1.26.16
whitenoise==6.5.0
